analyzer/pyproject.toml
[tool.poetry]
name = "analyzer"
version = "0.1.0"
description = "FIT file analysis tool for cycling and running workouts"
authors = ["Helge"]
readme = "README.md"
packages = [{include = "*", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.0.0"
numpy = "*"
pydantic = "^2.0.0"
fitparse = "^1.2.0"
pyyaml = "^6.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
fit-analyze = "fit_analyzes:main"
strava-analyze = "strava_analyze:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"